[tetrapak] add_directory doesn't remove old tasks
Summary: For a case, if it will be initialized twice, the tasks are not removed. The tasks were overwritten before. Now, it is possible to go with a normal procedure throw a normal register_tasks procedure. Register tasks merge the registered tasks before and new one. If there are tasks presented. There will be nothing happen.

Test Plan: make, read

Reviewers: orixmann, aschultz, vtarasen, hsychla

Reviewed By: hsychla

CC: mkraemer, aschultz

Differential Revision: https://review.tpip.net/D1245